feat(address-book): hide write controls for unauthenticated users
Split address-book page into a server wrapper (page.tsx) that resolves
auth session and a client component (AddressBookClient.tsx) that accepts
a canEdit prop.

Unauthenticated users see the full read-only address book but none of:
- Export JSON / Import JSON buttons
- + Add label button
- + Category / Edit buttons on rows
- Import error/success banners
- AddressLabelEditor modal

The API routes still enforce auth independently; this is a UX-layer change.
